split ForwardAnalysis into Core and Continuation
Similar to BackwardAnalysis, this contribution splits ForwardAnalysis
into two tasks, one which remains mostly identical to the old
ForwardAnalysis, and one which just continues an analysis with a
pre-occupied reachedSet. The algorithm employs the second type to pause
the analysis after a loop iteration (and request its continuation with
the separate task), in order to provide other tasks with an opportunity
to get picked for execution.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/block-based-task-partitioning@38983 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c