rework summary version management
Before this contribution, a situation could arise where a
ForwardAnalysis would abort because of another, more
recent ForwardAnalysis on the same block; which in turn could abort
because it would recognize the refined summary it depends on as
redundant. In this case, summary propagation would wrongly ceade.
The root cause for this were the limited capabilities of the
previous summary version management.

This contribution resolves this issue.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/block-based-task-partitioning@39002 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c