Copy branch with formula witness to orphaned branch 'alternative-violation-witness_formula-witness'
This branch contains an alternative violation witness format, called formula witness.
It is used with option 'analysis.useAlternativeErrorWitnesses=true'.

A formula witness only works with predicate analysis.
It is a directory that contains a file `path`
with the sequence of blocks (as ids) that lead to the target state,
and one file for each block which contains the abstraction (as SMTlib2 formula)
that was computed for this block.

The corresponding code is
in `src/org/sosy_lab/cpachecker/core/algorithm/alternative_error_witness`.


In the past, these formula witnesses were used in component-based CEGAR
[Beyer, Haltermann, Lemberger, Wehrheim at ICSE 2022]
to allow the predicate analysis to quickly rebuild the ARG
of a target path found in a previous predicate-analysis run
and check it for feasibility.

This reconstruction of a previous predicate-analysis ARG can be done
with 'cpa.predicate.refinement.useExternalErrorPath=true'
and 'cpa.predicate.refinement.formatWitness=PATH_TO_WITNESS/path'.




git-svn-id: https://svn.sosy-lab.org/software/cpachecker/orphaned/alternative-violation-witness_formula-witness@39140 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c