Use pako instead of zlib for decompressing the JS worker code
Pako is actually the JS port of zlib,
and the JS module zlib is just a wrapper with extended functionality
around pako.
Because we only need very little of it,
we can easily do this ourselves and avoid the need for zlib.

This improves things because zlib is actually not a regular npm module
(we never installed it), it is one of the modules that are provided by
default by node.
In order to make the code work in a browser,
webpack always automatically inserted a polyfill for zlib.
The next major version of webpack won't do that anymore,
so we would have to install the library,
but it is easier to just use pako directly.
And of course this reduces the bundle size a little bit.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@39286 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c