Implement the proposed fix for #941
With the following change, witness validation rejects the wrong invariant and reports "UNKNOWN".
- The checked loop-bound states are taken from the location where an invariant is specified
- The hypothesis includes all states within the checked range of BMC
- The violation includes all states not in the hypothesis

However, after the above revision, CTI extraction fails to find `badStateBlockingClauses`,
and hence initializing `InductionResult` throws an exception.
To make the revision work, `checkArgument()` in the constructor of `InductionResult` is temporarily disabled.
This problem needs further investigation.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/941-k-induction-unsound-validation@39574 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c