Simplify uses of FluentIterable just for calling uniqueIndex
There is no reason to go through a FluentIterable when the only use of
it is to call Maps.uniqueIndex().
Directly calling the latter makes the code better understandable
because it more clearly communicates that this is about maps
and it does not hide the main operation in the middle of the line.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@39624 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c