CFA for C programs supports GCC attributes (with focus on function declarations)
Because there's a vast amount of attributes and we do not want to
be selective during CFA creation (CPAs may find different attributes
useful), we do not parse attributes, but store them as a sequence of strings.

Many different language constructs can get attributes - we focus on
function declarations for now, because these are by far the most prominent
in our applications.

For general information on GNU C attributes, see
https://gcc.gnu.org/onlinedocs/gcc/Attribute-Syntax.html\#Attribute-Syntax

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/946-support-__noreturn__-or-a-selection-of-non-returning-functions@39646 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c