Support attribute 'weak'
The weak attribute causes a declaration of an external symbol to be emitted as a weak symbol rather than a global. This is primarily useful in defining library functions that can be overridden in user code, though it can also be used with non-function declarations. The overriding symbol must have the same type as the weak symbol. In addition, if it designates a variable it must also have the same size and alignment as the weak symbol. Weak symbols are supported for ELF targets, and also for a.out targets when using the GNU assembler and linker.

This is not relevant for CPAchecker,
so we support the attribute by ignoring.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/946-support-__noreturn__-or-a-selection-of-non-returning-functions@39668 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c