Support attribute 'returns_twice'
The returns_twice attribute tells the compiler that a function may return more than one time. The compiler ensures that all registers are dead before calling such a function and emits a warning about the variables that may be clobbered after the second return from the function. Examples of such functions are setjmp and vfork. The longjmp-like counterpart of such function, if any, might need to be marked with the noreturn attribute.

This is probably not relevant for CPAchecker, so we support
by ignoring.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/946-support-__noreturn__-or-a-selection-of-non-returning-functions@39669 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c