Remove an anonymous Function instance from CParserUtils
Putting code into proper methods instead of anonymous Function instances
is better to read and easier to test.
If the code is long enough that it should not be a lambda,
then it is also long enough to have its own function.

As a side effect from adjusting the callers,
let's also remove some reassignments to local variables.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@39673 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c