Replace static Function and Predicate fields with methods
There is almost never a reason to store Function and Predicate instances
in static fields as constants, it is almost always better to just use
regular methods. This makes the code easier to read and test.
In some of the cases changed here the instances were actually never even
used in a functional way, the code just called apply() on them manually.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@39676 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c