Reformat core packages of CPAchecker
This is the start of a reformat of all classes in CPAchecker
with our code formatter google-java-format.

More than 5 years ago we started using google-java-format,
but at that time decided to not reformat the whole code base
in order to not complicate the merge of open branches
and keep "svn blame" useful.
Instead we used the "diff mode" of google-java-format, which allowed us
to format only changed lines of a file, so at least new code
and changed code could be formatted while not touching existing code.

The goal was that the amount of properly formatted code would increase
over time, and we can now see that this has worked as hoped:
92% of CPAchecker's Java code is already properly formatted!
Thank you to all of you who contributed to this!
While there are still 1800 files out of 2700 Java files
that are not completely formatted with google-java-format,
there are only 37000 lines out of 470000 lines
that would be touched by the formatter.
And, of course, most of these lines will be lines that have not been
touched in the last five years, so reformatting them will be less
problematic than reformatting actively developed code.

Furthermore, while the "diff mode" of google-java-format is nice,
there are still two problems with it:
Sometimes google-java-format formats more than just the changed
lines, so either the commit would become larger than it should and more
difficult to review, or the developer had to put in manual effort to
commit only the really desired changes.
Furthermore, we couldn't enforce this properly via CI checks,
so it happened that new or changed code was committed
that was not properly formatted as it should be.

So because the percentage of unformatted lines has decreased
and thus the impact of a global reformat also decreased,
we believe that now has come the time where a reformat
and the advantage of not having to fiddle with "diff mode" anymore
are worth the hassle of potential merge conflicts and more indirections
when looking at blame information.

Not all files are formatted in this commit yet,
the reformatting is split over several commits
in the hope that this makes merging and conflict handling easier.
The final commit will update the documentation and CI checks
such that formatting changes and actual changes are not mixed in the
same commit.

If you have trouble merging these changes into your branch,
we suggest to merge step-wise by starting with merging all changes
from before the commit with "svn merge ^/trunk@39782"
and then going through the reformatting commits.
If there are still too many conflicts to handle,
please contact us on the mailing list,
we will attempt to assist you.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@39783 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c