Fix configuration bugs in VA loop summary config
The option cfa.strategies does not exist,
this needs to be cfa.summaries.strategies.
Also the options that are disabled by this commit
might lead to artificial differences when comparing to
valueAnalysis.properties, because this e.h. changes
traversal order etc. Some of them also do not make sense
as they are specific to predicate analysis, so I deleted
them right away.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/loopsummary@39802 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c