Rewrite uses of ImmutableMap.Builder.build() to buildOrThrow()
The two methods do the same, but the name of the latter is more clear
and build() will be deprecated by Guava soon.

Furthermore, update Guava because the newest version
provides buildKeepingLast() as alternative method
that does not throw on duplicate keys,
and we want to reference this method in our error message
of the respective CI job.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@39843 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c