Optimize "foo == foo" and similar conditions away in ASTConverter
The result of comparing an int variable against itself is always
statically known, so we can optimize such conditions away
already when creating the CFA, similarly to what we do for other
conditions that we know how to evaluate.

This optimization actually has the benefit of hiding
the bug from #940.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@39877 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c