Add check in InvariantsState.widen() that it does not return null
Doing so would cause a crash in InvariantsState.
Revision r27600 / 6458415f added the null check here, but
a) the null check would not help if the loop would perform
   another iteration after result became null,
a) there is only one caller of widen() and it would still crash, and
b) IMO it does not make sense for a state to become bottom during
   widening.

Cf. https://gitlab.com/sosy-lab/software/cpachecker/-/commit/6458415fc83f2312ad305a2c9a76ce8a43bf9b21#note_878652762
and #940 for more discussion of this.

So now we fail early with a descriptive error message.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@39880 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c