Improve CPAchecker's reaction to time limits and Ctrl+C
This is related to r38354 / e16549e7
and basically reverts that commit,
but additionally fixes the problem that that commit intended so solve,
without the unwanted other effect added there.

During the phase where we print statistics,
we do not want to terminate: if we have come that far,
we want to finish printing the statistics.
So before r38354 we would disable the code that reacts to external
signals like Ctrl+C and turns them into shutdown requests
before entering the statistics phase.
However, what we did not consider was that because we disabled our
shutdown hook completely, there was nothing anymore that prevented the
JVM from terminating immediately when receiving Ctrl+C.

So r38354 intended to improve that situation by keeping the shutdown
hook active during statistics. This would prevent the JVM from exiting,
but it would also mean that the shutdown hook would now send shutdown
requests during the statistics phase, which we also do not want.

So now we change ShutdownHook such that we can disable the shutdown
requests that it issues but still let it keep blocking the JVM
termination. Together with reverting r38354 this should finally give us
the desired behavior in all circumstances.

More information about this is also in the comments added by this commit and
https://gitlab.com/sosy-lab/software/cpachecker/-/commit/e16549e7ca8b42ec9bd2979c9a1a86101d8c8cf5#note_691312180

The only actual change here is the revert of r38354 in CPAMain
and splitting the "enable" flag in ShutdownHook into two fields.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40116 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c