Add first draft of abstract acceleration strategy
This already contains an algorithm that determines constraints
over CFAEdge visit frequency relations. Works for the simple
test program absacc1.c, but will probably break for nested loops
and function calls for example. The next step will be to
determine for the CFAEdges whether they increment the variables
and then we can construct the ghost CFA from that.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/loopsummary@40354 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c