Fix exception if not all irrelevant edges are removed before CFA construction in `SliceToCfaConversion`
During program slice to CFA conversion, irrelevant assume edges (i.e., edges that are not
part of the program slice) are not removed, because `CFASimplifier` (a CFA post-processor)
only works if assume edges are not replaced by blank edges (irrelevant edges are typically
replaced by no-op blank edges).

An AST node substitution for CFA edges is applied to all edges left in the program slice CFA.
It fails, because it expects all edges left in the program slice CFA to be relevant.
Adding a simple `if (pSlice.getRelevantEdges().contains(edge))` to the substitution fixes this issue.

Fixes #965

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/965-slice-exporter-irrelevant-edge-exception@40362 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c