Don't store references to function exit nodes that are not part of the CFA
Some function entry nodes do not have a corresponding function exit node that is also part of
the CFA . If a function never returns, because it always aborts the program or always executes
an infinite loop, the CFA doesn't contain an exit node for the function.

Before this commit, instances of `FunctionEntryNode` could contain references to
function exit nodes that were unreachable and not part of the CFA.
This commit makes the `exitNode` field in `FunctionEntryNode` nullable and
`FunctionEntryNode#getExitNode()` return an `Optional`. This makes it explicit that
some function entry nodes do not have a corresponding function exit node that is also part of
the CFA. You really have to think about this special case now.

During CFA construction, function exit nodes (even unreachable ones) are created but later removed
if they don't have any entering edges. This made implementation a bit easier, because currently
you don't really know at the time when function entry and exit nodes are created, whether entering
edges will be added to those exit nodes.

Fixes #951

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/951-dont-store-references-to-unreachable-function-exit-nodes@40379 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c