Fix single-loop-transformation in case of empty infinite loops
Fixes #956
Fixes #960

The case "if (x) { label; goto label; }" was not handled correctly
and lead to non-termination of the single-loop transformation,
eventually resulting in an OOM crash.
r40445 (8556668b) avoided the non-termination but left loops like the
above unhandled, i.e., they were not removed.
This led to a NullPointerException later on.

What we have to do instead is to treat these loops just like
"while (x) {}" loops and replace them with a termination node.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/cfa-single-loop-transformation@40474 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c