Weaker type for parameter of ARGUtils.getPathFromBranchingInformation
This method needs access to the branching information
(which branch to take) on each ARG branching,
but we do not need to force callers to present this as a map from the
id of the ARGState to the result.
Using a function with parameter type ARGState is more flexible for
callers.
And then we can also pass the AssumeEdge that we are interested in
because it might be useful and we have it anyway.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40584 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c