Add a replacement for the "branching formulas" of the predicate analysis
Part of #971

With predicate analysis, in many cases we have an ARG that is a DAG
and in the end we need to find a single feasible path from the root
state to a target state in this DAG.
We do so by creating a boolean SMT variable for each branching point in
the ARG and add clauses that force a specific value for each of these
variables depending on whether the respective left or right branch
should be taken. Then we solve this (incrementally on top of the query
that encodes reachability of the ARG states), get the variable values
from the solver, and compute the ARG path from this information.

Creating and solving these additional clauses is cheap,
because they do not cause new conflicts in the solver,
so the solver only needs to take its existing proof of satisfiability
and compute the assignment of values for the new variables that follow
from the existing proof.
However, we found cases where the number of additional variables
gets so large that the time for creating a model for them is excessive.
For example, for BMC, k=8, and eca-programs/Problem101_label11.c from
SV-Benchmarks the actual analyses finishes in <10min,
there are >500000 branching variables, and
creating and solving the branching formula still takes just a few seconds,
but creating a model for it takes an hour.

The idea of the new solution is that instead of creating a branching
formula with information about each branching in the ARG,
we ask the solver only about those branchings that we encounter
when computing the path from root to target.
And we do not even need to create new variables for it
nor do we need to ask the solver for complete model enumeration.
Instead we can directly ask the solver to evaluate those clauses
that we need for each branching point.
In the example above, the whole counterexample analysis
is finished after a few seconds.

In this commit, we only add the new method and mark the old one as
deprecated, but do not use it yet.
Callers will be added in the next commit.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40585 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c