Avoid use of branching formulas for BMC and related algorithms
Part of #971

As explained in the previous commit,
branching formulas can be expensive
and we now have a new solution that avoids their cost.

This commit lets AbstractBMCAlgorithm and PdrAlgorithm make use of it.
This improves the performance of counterexample analysis for our
implementations of BMC, k-induction, IMC, ISMC, and PDR.
Note that this improvement even applies if all output files are disabled.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40586 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c