Disable fading of rendering-state modal in HTML report
Fixes #972

On very fast machines, the rendering of the graph is faster than the
time the modal takes to be shown, because a fading effect is applied for
the latter.
And it seems that if we tell the modal to hide while it is still fading
it, Bootstrap ignores this
(cf. https://github.com/twbs/bootstrap/issues/30584).
More explanations are in the linked issue.

Because we do not need any animation for the modal,
the easiest fix is to remove the fading.
Now hiding and showing of the modal should be synchronous.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40598 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c