Fix bug in formula handling in PredicateCPAInvariantsManager
r26447 (9ec43f09) introduced BlockFormulas
and r27161 (a1f822df) made it immutable with a defensive copy,
but in PredicateCPAInvariantsManager there was code that kept modifying
the list of formulas afterwards and relying on this to have an effect.
This is of course quite ugly and was not noticed when BlockFormulas
became immutable.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40628 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c