Avoid use of branching formulas for CEGAR-based predicate analysis
Part of #971

As explained in the issue and in r40585 (b8e0c263),
our branching formulas can be expensive
and we now have a new solution that works without them.

This actually adds a layering violation regarding the hierarchy of
classes and responsibilities that we currently have in the predicate
analysis. Code that directly works with the solver is usually fully
decoupled from high-level concepts like the ARG, abstract states, etc.
Now we have code that handles ARGStates and the solver model at the same
time. Furthermore, this part of the predicate analysis gets coupled to
the ARGCPA.

However, there are the following arguments in favor of this solution:
- The better efficiency.
- It might actually be easier to understand despite the layering
  violation because now we have exactly one place that determines the
  error path from a solved list of formulas. Previously this was spread
  over several classes (creating of branching formula, solving of
  branching formula, interpreting the model, mapping it to the ARG)
  and one had to understand the idea of our branching formula as well.
- This remains optional,
  all predicate-related code can still be used without ARGStates.

There is one existing piece of code whose functionality is not
reimplemented with the new solution:
r26447 (9ec43f09) introduced the possibility to override the formulas
for assume edges that are used while analyzing the branching,
and such overriding was added to BAMPredicateRefiner in the same commit:
https://gitlab.com/sosy-lab/software/cpachecker/-/commit/9ec43f091e916787e2900f186e69e63d6f81874b#236e23bb0a2338db44ee73a17730e6e8e4d56507_200_204
However, it is unclear why this was added:
The commit message does not explain anything and the "TODO" comment in
the JavaDoc for this parameter was never replaced with something
meaningful.
BAM itself is considerably older than that change and seems to have
worked without it. The integration tests for BAM also do not show a
difference.

Furthermore, an additional small change is that imprecise
counterexamples (those were computing the precise path failed) are not
associated with a model of the path formula anymore,
which means that no variable values are present in the output.
This should have little effect because by default we do not dump
imprecise counterexamples at all, and if we do, we provide the SMT
formula from which users could get the model using an external SMT
solver if they really need it for debugging.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40634 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c