See #928: Fix two bugs in label handling
Before, a label followed by an if statement caused a problem,
because the edges did not overlap until the condition of the if.
Also, labels followed by a CFAEdge without a real file location
caused an exception during the computation of the next real file
location.

The two new example programs show, that these bugs have been fixed.
However, if-else_with-goto-out-and-back-in.c has differently
ordered statements in export (with different control flow!) because
it is not possible yet to compute the original end of the then-
branch.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/928-CFA-to-C-exporter-prototype@40636 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c