Fix counterexample analysis for BAM and replace "branching formula"
r40634 (0827db0c) broke counterexamples for BAM-based analyses
because the feature that was discussed in that commit's message
was actually used (just not in the integration tests).
So now we use it again for the new implementation.
(We just need to reconnect the pieces actually,
because the code had not been removed yet.)

This commit also removes the previous "branching formula"
and is part of #971.
Because we still require the above feature for BAM,
we replace it with the possibility to provide a map of "branching formulas",
which should not be confused with the previous "branching formula".
The latter was a (potentially large) formula with a (potentially large)
number of variables.
The new "branching formulas" are just small formulas representing the
expressions of assume edges in the ARG.
These formulas are what the code in
PathFormulaManager.getARGPathFromModel() would usually recreate anyway,
but for BAM we provide a way to overwrite them.
But the performance improvements and simplifications that were gained
from avoiding the use of the previous "branching formula" all still
exist.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40642 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c