See #928: Change approach and use AST instead of CFA information
Until now, we have used the file location and the raw statement of
the CFAEdges for the export.

Looking at the two new example programs, that are added with this
commit, one can see that this information does not suffice for the
export.
In if-else_with-goto-out-of-both-branches.c, both branches of the
if-statement end with a jump out of the if-statement to a
respective return-statement.
The other program, if-else_with-labeled-return-within.c, has the
same control-flow. However, the labeled return-statements are
located within the respective branch.
The problem is that the resulting CFAs are the same. In both cases,
the MergePoint is the FunctionExitNode, but what we are interested
in in export is the location of the respective closing brace of the
branches.

In special cases, it is possible to find out where the closing
brace must have been.
For example, the file location of the last edge belonging to the
original then-branch has to be smaller than the file location of
the first edge of the else-branch. That means that an edge in the
then-branch of the CFA that has a bigger file location than the
first edge of the else-branch must be located outside of the
original if-statement.

However, there are cases where the information does not suffice to
decide whether an edge was originally part of the if-statement.
E.g., there is no way to differentiate when there is a jump out of
the else-branch to a label that directly follows the if-statement
compared to the program where the label directly follows the goto-
statement (meaning that the labeled statement is still located
within the if-statement).

For this reason, we need encapsulation information to stay closer
to the original program in export.
This is provided in the AST of the original program.
As the Eclipse CDT parser provides all the information we need for
the identity export (meaning without CFA transformations), we can
simply use the resulting AST's getRawSignature-method for export.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/928-CFA-to-C-exporter-prototype@40651 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c