Reactoring: Make getPathFromBranchingInformation not care about path end
Previously, it checked that the produced path ends in a target state,
and this behavior could be disabled by a boolean argument
(but this possibility was never used).
Now we have only three callers and for one of them we actually want a
different check anyway, so it is cleaner to remove this check from the
utility method and put it into the callers.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40652 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c