Remove option cpa.predicate.refinement.discardErrorPath
This was added because IMC and ISMC reuse existing code for analyzing
counterexamples but do not care about actual counterexamples, only about
interpolants. So we do not want to spend useless effort.

Since the recent removal of branching formulas in r40634 (0827db0c)
the respective method in InterpolationManager anyway only spends time
for analyzing counterexamples if an ARGPath is passed in
(because it needs this to start with) and IMC and ISMC do not do this.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40654 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c