Provide method InterpolationManager.interpolate()
A method doing this was basically already present and used,
but both parameters and return values were wrapped into other objects
that are useless for this purpose, and the method was confusingly named
(buildCounterexampleTrace).
The new API is much more clear for those callers that do not need a
counterexample trace.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40655 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c