Restructure public API of PathChecker
There were two public methods without a clear distinction for callers
which to use (one was even undocumented).
But actually for callers of the second method (createCounterexample)
it doesn't matter, they can also call the first method
(handleFeasibleCounterexample), which is just a thin wrapper around the
second. So let's do this and make the second private.

The only change is that now the option alwaysUseImpreciseCounterexamples
also has some (limited) effect for BMC, not only for CEGAR-based
predicate analysis. This is even better than the option having no effect
for some configurations.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40656 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c