Improve time statistics for counterexample handling of predicate analysis
Basically, we have two steps in counterexample handling:
First, we determine the actual error path in the ARG from the model,
and then we re-check the error path to get a variable assignment
(this is what PathChecker does).
Now we have separate statistics for both steps,
and consistently for BMC-based and CEGAR-based analyses
with similar names for the two steps
(error-path creation and error-path post-processing, respectively).

Furthermore, the sat-check time of InterpolationManager missed solver
calls in the fallback path if interpolation failed.
This is fixed as well.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40659 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c