Simplification for PathFormulaManager.getARGPathFromModel
It makes sense to provide one method
with the full set of possible parameters
and one method with the minimal set of required parameters
to make it convenient for callers in simple cases.
But we do not need to provide a method with an intermediate number of
parameters because all existing callers can use the simple version
anyway.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40661 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c