Make procedure attributes mandatory for all SDG nodes
They weren't mandatory because we want our implementation to also represent
program dependence graphs (PDGs) that are only for a single procedure,
so the procedure attribute would always be the same. There is no need
to specify the same procedure for every node, so we made it optional.

This made handling procedures inconvenient, because we always had to check
whether the procedure is actually present.

This is why this commit makes procedure attributes mandatory.
For PDGs the procedure type `Void` can be used.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/sdg-refactoring@40771 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c