Add `SdgEdge` to store the cause variable in SDG edges
Without the new `SdgEdge` class, we cannot determine whether a flow
dependency for a specific variable exists. We can only look at what
variables a node defines and uses, but this is not always accurate
(e.g., if there are two variables, we don't know which one or
whether both variables caused the flow dependency).

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/sdg-refactoring@40775 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c