Disable traversal by distance to target in predicate cex checks
Currently all configs for witness validations and counterexample checks
have an iteration order that is not reverse post-order,
but orders states by distance to the target state.
This is highly inefficient for all analyses that do not use merge-sep
but might merge states (because one path will be followed,
and then other paths may be found and followed,
which causes merges and re-analysis of the previous paths).

For predicate analysis with CEGAR as counterexample check
this can even cause crashes,
because the predicate refiner does not expect the ARG shapes
that result from a non-reverse-post-order traversal strategy.

This is discussed in #907.
Until benchmark results for witness validation exist,
we at least disable this for counterexample checks with predicate
analysis.
This produces for example a considerable speed up and less crashes
for -predicateAnalysis-slicing.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40787 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c