Remove label-adder implementation because it's unused
This was used for a very short time to add test-goal labels in programs,
for example for branch- and function coverage.
But it was soon replaced by own implementations based on pycparser [1]
and clang libtooling [2].
In addition, instrumenting C programs with C labels does not fit the
main focus of CPAchecker, so there's little probability that this class
will be needed in the future.

[1]: https://gitlab.com/sosy-lab/software/conditional-testing/-/blob/4f0005fb745c3ae2b1e65ecf4b837100d4ea888d/condtest/instrumenter.py
[2]: https://gitlab.com/sosy-lab/software/test-suite-validator/-/tree/b0b23fd531749b32ae0e22b5330e05e81deef015/label-adder

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40825 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c