Fix handling of UFs in AssignmentToPathAllocator.createAssignments
The current code adds values for (non-nullary) UFs
to the map of variables, which does not make sense
because UFs do not correspond to program variables.
We use UFs for stuff like modeling the heap memory.
Furthermore, the current code ignores the UF arguments in the map key,
so in case there exist the model assignments
"foo(0) = 0" and "foo(1) = 1" it would just store one of the two values
for "foo".

This seems to have no negative effects because the values would be
ignored in the map of variables later, but it still is better to avoid
these wrong terms.

So now we only store variable values in the map of variables
and values for non-nullary UFs in the map of functions.

This was also the behavior before r16931 / 19e7bcb2.
That commit removed the "if (term instanceof Variable)"
and made the code for handling variables apply to both
variables and non-nullary UFs.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40828 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c