Fix calls to PTS.isBaseName from r25642 / b3605788
Since this commit, we need to call FormulaManagerView.parseName()
first before calling PointerTargetSet.getBase().
Technically, it works if we call PointerTargetSet.isBaseName()
without calling FormulaManagerView.parseName(),
but this is just a coincidence due to how FormulaManagerView
and PointerTargetSet mangle variable names in a different manner.

So we apply the proper name handling now.
There are only these two callers of isBaseName().

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40829 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c