Add a workaround for missing reverse post-order id to CFASecondPassBuilder
As discussed in #989, CFASecondPassBuilder creates additional CFA nodes
that miss the reverse post-order id.
This can cause all sorts of problems, from inefficient CFA traversal
to crashes in predicate analysis because of unexpected ARG branching.
The correct fix is likely to compute the reverse post-order ids
after CFASecondPassBuilder runs, but this is more complex.

As a temporary workaround and in order to be able to see whether this
bug had visible effects, we just set a reverse post-order id that is
less wrong than the default: we copy the id from the predecessor node.
This still does not guarantee unique ids,
but at least analyses that use the ids just for the traversal order
should now work fine.

This change should be reverted once #989 is properly fixed.

git-svn-id: https://svn.sosy-lab.org/software/cpachecker/trunk@40836 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c